[Quanta/platform] pmbus driver not include kernel source tree and use it directly
